+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   previously dumped to the file system.
 - Method (`DataFrameDescriber.format_table`) to format the dataframe's table
   using `Table`.
-
+- CLI feature to write formatted tables using YAML and CSV files as an Excel
+  file.
 
 ## [0.1.0] - 2023-08-16
 Downstream moderate risk update release.